requirements.txt
Pillow == 8.0.1
cogwatch == 2.0.0
discord.py == 1.6.0
python_dotenv == 0.15.0
requests == 2.25.1
youtube_dl == 2021.1.24.1
